ISSUE-13 Support migration on streaming aggregation state format 1 to 2 (#16)
This patch introduces `StreamingAggregationMigrator` to help migration streaming aggregation state from 1 to 2. Here's the target:

* query used streaming aggregation (mostly `agg()`)
* the query ran before Spark 2.4.0 for the first time, and checkpoint is kept

Schema of state key/value should be provided, so I don't create CLI app for this. When Spark will contain state schema, we can leverage it and finally make this to CLI app.

Note that it migrates state for latest committed batch and it's not done per batch. You can't roll back to previous batch and need to start with latest committed batch. (It will be picked up from Spark automatically.)

closes #13