ISSUE-14 Support migration on flatMapGroupsWithState state format 1 to 2 (#19)
This patch introduces `FlatMapGroupsWithStateMigrator` to help migration state which is from flatMapGroupsWithState from version 1 to 2. Here's the target:

* query used flatMapGroupsWithState
* the query ran before Spark 2.4.0 for the first time, and checkpoint is kept

Schema of state key/value should be provided, so I don't create CLI app for this. When Spark will contain state schema, we can leverage it and finally make this to CLI app.

Note that it migrates state for latest committed batch and it's not done per batch. You can't roll back to previous batch and need to start with latest committed batch. (It will be picked up from Spark automatically.)

This patch also sorts out README.md, but we may need to sort out once again to let users just follow like HOW-TO. (Currently it focuses to show all features one by one.)

closes #14